python-encryption/requirements.txt
Flask
PyPDF2
reportlab
cryptography
pycryptodome